Removing Bcrypt gem which resulted in a few other changes
Edited the db for the users model so that it has a password field in place of the password_digest field. Also made some other changes to make it so that rails tests functions properly, such as replacing name with username. Finally, I also added one line of code in the routes file so that the sign up page functions properly.